requirements.txt
fastapi
uvicorn
pydantic
pyttsx3
langchain-community
PyMuPDF
groq
sqlalchemy
opencv-python
python-multipart
dotenv
openai
faiss-cpu
requests
passlib[bcrypt]
PyJWT